requirements.txt
# Dependencias directas de savagePy
# Versiones exactas del entorno de desarrollo (pip freeze Feb 2026)

# Servidor web
Flask==3.1.3

# HTTP / cliente NocoDB
requests==2.32.5

# Generación de PDF
weasyprint==68.1

# Generación de Word (.docx)
python-docx==1.2.0
docxtpl==0.20.2
pypandoc==1.15        # conversión Markdown → Word para el compendio de reglas

# Variables de entorno (.env)
python-dotenv==1.2.1

# Nota: el resto de paquetes del freeze son dependencias transitivas
# que pip instala automáticamente al instalar los de arriba.
#
# Instalación:
#   python -m venv .venv
#   .venv/Scripts/activate        # Windows
#   source .venv/bin/activate     # Linux/Mac
#   pip install -r requirements.txt
#
# WeasyPrint en WSL requiere además:
#   sudo apt install libpango-1.0-0 libpangoft2-1.0-0 libharfbuzz0b
#
# pypandoc requiere pandoc instalado en el sistema:
#   sudo apt install pandoc
